# docs/api/OPENAPI.yaml
openapi: 3.0.3
info:
  title: PEZHWAN Identity Server
  description: |
    Universal Identity & Access Management API. Reference host for the
    `@pezhwan/*` SDK stack. All endpoints use a `{ success, data, error }`
    envelope except OAuth/OIDC, which follow OAuth 2.1 / OIDC wire format.
  version: 0.1.0
servers:
  - url: http://localhost:4011
    description: Development reference server
security:
  - {}
...
paths:
...
  /v1/sessions:
    get:
      tags: [Sessions]
      security:
        - Bearer: []
      summary: List the user's active sessions
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
...
  /v1/mfa/verify:
    post:
      tags: [MFA]
      security:
        - Bearer: []
      summary: Verify a TOTP/backup code (step-up)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
              required: [code]
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean }
...
  /v1/oauth/token:
    post:
      tags: [OAuth]
      summary: Token endpoint (code exchange, refresh, client_credentials)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token, client_credentials]
                client_id: { type: string }
                client_secret: { type: string }
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string }
                refresh_token: { type: string }
                scope: { type: string }
              required: [grant_type, client_id]
      responses:
        '200':
          description: Token response (RFC 6749)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthToken' }
        '400':
          description: invalid_request / invalid_grant / unsupported_grant_type
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  error_description: { type: string }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Access token from /v1/auth/login|register|refresh (JWT RS256)
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: API key (pk_live_...) as `Bearer <key>`
    RoleAdmin:
      type: http
      scheme: bearer
      description: Requires the ADMIN role on the identity
  schemas:
...
    Session:
      type: object
      properties:
        _id: { type: string }
        userAgent: { type: string, nullable: true }
        ip: { type: string, nullable: true }
        createdAt: { type: string }
        lastSeenAt: { type: string }
    OAuthToken:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }
        refresh_token: { type: string }
        id_token: { type: string }
        scope: { type: string }
...